server/lib/python/cartodb_geocoder/requirements.txt
redis==2.10.5
python-dateutil==2.4.2

# Test
mock==1.3.0
mockredispy==2.9.0.11
nose==1.3.7
